fixed unit test failures

+++ services/delete_totalmobile_jobs_service.py
@@ -1,4 +1,3 @@
-from models.totalmobile.totalmobile_jobs_model import TotalmobileJobsModel
 from services.blaise_service import BlaiseService
 from services.totalmobile_service import TotalmobileService
 
@@ -12,22 +11,14 @@ class DeleteTotalmobileJobsService:
         self.world_id = "13013122-d69f-4d6b-gu1d-721f190c4479"
 
     def delete_totalmobile_jobs_completed_in_blaise(self) -> None:
-        totalmobile_job_references = self.get_incomplete_totalmobile_job_references(self.world_id)
-        totalmobile_jobs_model = TotalmobileJobsModel(totalmobile_job_references)
+        totalmobile_jobs_model = self.totalmobile_service.get_jobs_model(self.world_id)
 
         for questionnaire_name in totalmobile_jobs_model.questionnaire_jobs.keys():
             completed_case_ids = self.get_completed_blaise_cases(questionnaire_name)
 
             for job in totalmobile_jobs_model.questionnaire_jobs[questionnaire_name]:
-                if job.case_id in completed_case_ids:
+                if job.visit_complete is False and job.case_id in completed_case_ids:
                     self.totalmobile_service.delete_job(self.world_id, job.reference, "110")
-
-    def delete_totalmobile_job(self, world_id: str, job_reference: str, reason: str) -> None:
-        self.totalmobile_service.delete_job(world_id, job_reference, reason)
-
-    def get_incomplete_totalmobile_job_references(self, world_id: str) -> list[str]:
-        jobs = self.totalmobile_service.get_jobs(world_id)
-        return [job["identity"]["reference"] for job in jobs if not job["visitComplete"]]
 
     def get_completed_blaise_cases(self, questionnaire_name) -> list[str]:
         case_status_list = self.blaise_service.get_case_status_information(questionnaire_name)
